Moved iteration strategy tests to different test files
